+++ data-raw/penguins_stats.R
@@ -12,6 +12,7 @@ penguins_stats <- penguins_dt[, .(
   avg_body_mass = mean(body_mass_g),
   min_body_mass = min(body_mass_g),
   max_body_mass = max(body_mass_g)
-), by = species]
+), by = species] |>
+_[, mid_labels := as.character(round(df$avg_body_mass))]
 
 usethis::use_data(penguins_stats, overwrite = TRUE)